requirements.txt
pandas>=1.4
numpy>=1.22
scikit-learn>=1.0
pyyaml>=6.0
typer>=0.9
matplotlib>=3.5
